Reorganize TypeScript definitions as well
