--- src/types/core/index.d.ts
@@ -1,19 +1,0 @@
-export { annotate, Annotation } from './Annotation';
-
-export {
-    andThen,
-    dispatch,
-    err,
-    errValue,
-    expect,
-    isErr,
-    isOk,
-    map,
-    mapError,
-    ok,
-    Result,
-    toString,
-    unwrap,
-    value,
-    withDefault,
-} from './Result';
