+++ src/types/format/index.d.ts
@@ -1,4 +1,2 @@
-import { Annotation } from '../../annotate';
-
-export function serialize(ann: Annotation): string;
-export function summarize(ann: Annotation): string;
+export { serialize } from './serialize';
+export { summarize } from './summarize';
