+++ src/types/format/serialize.d.ts
@@ -1,0 +1,3 @@
+import { Annotation } from '../annotate';
+
+export function serialize(ann: Annotation): string;
