+++ src/types/format/summarize.d.ts
@@ -1,0 +1,3 @@
+import { Annotation } from '../annotate';
+
+export function summarize(ann: Annotation): string;
