+++ src/types/result.d.ts
