+++ src/types/stdlib/boolean.d.ts
@@ -1,4 +1,4 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export const boolean: Decoder<boolean>;
 export const truthy: Decoder<boolean>;
