+++ src/types/stdlib/date.d.ts
@@ -1,4 +1,4 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export const date: Decoder<Date>;
 export const iso8601: Decoder<Date>;
