+++ src/types/stdlib/describe.d.ts
@@ -1,3 +1,3 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export function describe<T>(decoder: Decoder<T>, msg: string): Decoder<T>;
