+++ src/types/stdlib/dispatch.d.ts
@@ -1,8 +1,8 @@
-import { Decoder, $DecoderType } from './types';
+import { Decoder, DecoderType } from '../_types';
 
 export type $Values<T extends object> = T[keyof T];
 
 export function dispatch<O extends { [key: string]: Decoder<any> }>(
     field: string,
     mapping: O,
-): Decoder<$Values<{ [key in keyof O]: $DecoderType<O[key]> }>>;
+): Decoder<$Values<{ [key in keyof O]: DecoderType<O[key]> }>>;
