+++ src/types/stdlib/either-tests.ts
