+++ src/types/stdlib/either.d.ts
@@ -1,4 +1,4 @@
-import { Decoder, Scalar } from './types';
+import { Decoder, Scalar } from '../_types';
 
 export function either<T1, T2>(d1: Decoder<T1>, d2: Decoder<T2>): Decoder<T1 | T2>;
 export function either2<T1, T2>(d1: Decoder<T1>, d2: Decoder<T2>): Decoder<T1 | T2>;
