+++ src/types/stdlib/fail.d.ts
@@ -1,3 +1,3 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export function fail(msg: string): Decoder<never>;
