+++ src/types/stdlib/instanceOf-tests.ts
