+++ src/types/stdlib/instanceOf.d.ts
@@ -1,3 +1,3 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export function instanceOf<T>(klass: new (...args: any) => T): Decoder<T>;
