+++ src/types/stdlib/lazy.d.ts
@@ -1,3 +1,3 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export function lazy<T>(decoderFn: () => Decoder<T>): Decoder<T>;
