+++ src/types/stdlib/mapping.d.ts
@@ -1,4 +1,4 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export function mapping<T>(decoder: Decoder<T>): Decoder<Map<string, T>>;
 export function dict<T>(decoder: Decoder<T>): Decoder<{ [key: string]: T }>;
