+++ src/types/stdlib/number.d.ts
@@ -1,4 +1,4 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export const integer: Decoder<number>;
 export const number: Decoder<number>;
