+++ src/types/stdlib/optional.d.ts
@@ -1,4 +1,4 @@
-import { Decoder } from './types';
+import { Decoder } from '../_types';
 
 export function optional<T>(decoder: Decoder<T>): Decoder<T | undefined>;
 export function nullable<T>(decoder: Decoder<T>): Decoder<T | null>;
